Add User Extensions, fix comments, Use new class ForgotPasswordEmailModel in forgotPassword
